Summarize the data

summary() is an analysis function that computes the protein summary statistics for a given tidyproteomics data object. This is a passthrough function as it returns the original tidyproteomics data-object.

Usage

summary(
  data,
  by = c("global"),
  destination = c("print", "save", "return"),
  contamination = NULL
)

Arguments

data

tidyproteomics data object

by

what to summarize

destination

character string, one of (save, print)

contamination

as character string

Value

a tibble on print, a tidyproteomics data-object on save

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# a global summary
ecoli_proteins %>% summary()
#> 
#> ── Summary: global ──
#> 
#>  files proteins protein_groups peptides peptides_unique quantifiable       CVs
#>      8     2772           2772    31271           31025         98.5 0.1573717
#> 

# a summary by sample
ecoli_proteins %>% summary("sample")
#> 
#> ── Summary: sample ──
#> 
#>  sample files proteins protein_groups peptides peptides_unique quantifiable
#>      ko     4     2772           2772    31271           31025         98.4
#>      wt     4     2772           2772    31271           31025         98.5
#>        CVs
#>  0.1380689
#>  0.1612074
#> 

# a summary by sample with match_between_runs removed
ecoli_proteins %>%
   subset(match_between_runs == FALSE) %>%
   summary("sample")
#> 
#> ℹ Subsetting data: match_between_runs == FALSE
#> ✔ Subsetting data: match_between_runs == FALSE ... done
#> 
#> 
#> ── Summary: sample ──
#> 
#>  sample files proteins protein_groups peptides peptides_unique quantifiable
#>      ko     4     2683           2683    30779           30562         99.1
#>      wt     4     2666           2666    30811           30589         99.1
#>        CVs
#>  0.1329208
#>  0.1558694
#> 

# a summary of match_between_runs
ecoli_proteins %>% summary('match_between_runs')
#> 
#> ── Summary: match_between_runs ──
#> 
#>  match_between_runs files proteins protein_groups peptides peptides_unique
#>               FALSE     8     2772           2772    30795           30576
#>                TRUE     8      527            527      477             450
#>  quantifiable       CVs
#>          99.1 0.1532708
#>          55.7 0.2856872
#> 

ecoli_proteins %>% summary("cellular_component")
#> 
#> ── Summary: cellular_component ──
#> 
#>            cellular_component files proteins protein_groups peptides
#>                  cytoskeleton     8     2772           2772    31271
#>                       cytosol     8      589            589     8894
#>  non-structural extracellular     8        3              3        9
#>                       nucleus     8        1              1        5
#>          other cell component     8      433            433     4684
#>   other cytoplasmic organelle     8     2772           2772    31271
#>               other membranes     8       61             61      744
#>               plasma membrane     8     2772           2772    31271
#>       translational apparatus     8        3              3       11
#>  peptides_unique quantifiable       CVs
#>            31025         98.5 0.1573717
#>             8828         99.1 0.1329716
#>                9         88.9 0.1248169
#>                5        100.0 0.1301963
#>             4652         98.6 0.1632274
#>            31025         98.5 0.1573717
#>              739         97.5 0.1871234
#>            31025         98.5 0.1573717
#>               11        100.0 0.4519613
#> 

ecoli_proteins %>% summary("biological_process")
#> 
#> ── Summary: biological_process ──
#> 
#>                biological_process files proteins protein_groups peptides
#>                    DNA metabolism     8     2772           2772    31271
#>   RNA metabolism OR transcription     8     2772           2772    31271
#>                     cell adhesion     8        1              1        1
#>  cell cycle OR cell proliferation     8        3              3       39
#>  cell organization and biogenesis     8       15             15      198
#>               cell-cell signaling     8     2772           2772    31271
#>           developmental processes     8        3              3        4
#>        other biological processes     8      175            175     1262
#>         other metabolic processes     8      610            610     8233
#>                protein metabolism     8       18             18      332
#>               signal transduction     8        1              1       18
#>                   stress response     8       45             45      330
#>                         transport     8      196            196     1542
#>  peptides_unique quantifiable       CVs
#>            31025         98.5 0.1573717
#>            31025         98.5 0.1573717
#>                1         62.5 0.2738303
#>               39        100.0 0.1084535
#>              197         98.4 0.1615593
#>            31025         98.5 0.1573717
#>                4         75.0 0.1980664
#>             1258         98.4 0.1752463
#>             8179         98.6 0.1511318
#>              332         99.9 0.1944947
#>               18        100.0 0.1228771
#>              316         93.6 0.1537054
#>             1529         97.6 0.1815972
#>